Put create_trainer_job and setup_deployment in a separate pyproject
