+++ fireworks-training-infra/fireworks_training_infra/__init__.py
@@ -1,0 +1,64 @@
+"""Provisioning helpers for Fireworks training jobs and inference deployments.
+
+Top-level re-exports for the common entry points; submodules
+(:mod:`fireworks_training_infra.infra`, :mod:`fireworks_training_infra.client`,
+:mod:`fireworks_training_infra.config`,
+:mod:`fireworks_training_infra.training_shapes`) are also available for direct
+import.
+"""
+
+from fireworks_training_infra.client import (
+    DCP_TIMEOUT_S,
+    DEFAULT_TIMEOUT_S,
+    GradAccNormalization,
+    ReconnectableClient,
+)
+from fireworks_training_infra.config import (
+    DeployConfig,
+    InfraConfig,
+    WeightSyncScope,
+)
+from fireworks_training_infra.infra import (
+    Infra,
+    ResourceCleanup,
+    StatusCallback,
+    TrainerHandle,
+    create_trainer_job,
+    get_deployment_gpu_count,
+    read_api_extra_headers_env,
+    request_deployment,
+    request_trainer_job,
+    setup_deployment,
+    setup_infra,
+    setup_or_reattach_deployment,
+    setup_training_client,
+    wait_deployment,
+    wait_trainer_job,
+)
+from fireworks_training_infra.training_shapes import auto_select_training_shape
+
+__all__ = [
+    "DCP_TIMEOUT_S",
+    "DEFAULT_TIMEOUT_S",
+    "DeployConfig",
+    "GradAccNormalization",
+    "Infra",
+    "InfraConfig",
+    "ReconnectableClient",
+    "ResourceCleanup",
+    "StatusCallback",
+    "TrainerHandle",
+    "WeightSyncScope",
+    "auto_select_training_shape",
+    "create_trainer_job",
+    "get_deployment_gpu_count",
+    "read_api_extra_headers_env",
+    "request_deployment",
+    "request_trainer_job",
+    "setup_deployment",
+    "setup_infra",
+    "setup_or_reattach_deployment",
+    "setup_training_client",
+    "wait_deployment",
+    "wait_trainer_job",
+]
