+++ fireworks-training-infra/fireworks_training_infra/training_shapes.py
