+++ training/pyproject.toml
@@ -8,6 +8,8 @@ version = "0.1.0"
 requires-python = ">=3.11"
 dependencies = [
     "fireworks-ai[training]>=1.2.0a70,<2",
+    "fireworks-training-infra>=0.1.0",
+    "tinker==0.15.0",
     "tqdm",
     "torch",
     "datasets",
